p3 done

+++ src/main/java/ru/javawebinar/topjava/repository/mock/InMemoryMealRepositoryImpl.java
@@ -1,17 +1,27 @@
 package ru.javawebinar.topjava.repository.mock;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import ru.javawebinar.topjava.AuthorizedUser;
 import ru.javawebinar.topjava.model.Meal;
+import ru.javawebinar.topjava.model.User;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
 
+import java.time.LocalDateTime;
 import java.util.Collection;
+import java.util.Comparator;
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.stream.Collectors;
 
 public class InMemoryMealRepositoryImpl implements MealRepository {
     private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
     private AtomicInteger counter = new AtomicInteger(0);
+
+    private static final Logger log = LoggerFactory.getLogger(InMemoryUserRepositoryImpl.class);
 
     {
         MealsUtil.MEALS.forEach(this::save);
@@ -28,19 +38,42 @@ public class InMemoryMealRepositoryImpl implements MealRepository {
         return repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal);
     }
 
+    // false if not found or don't have permission
     @Override
-    public void delete(int id) {
-        repository.remove(id);
+    public boolean delete(int id) {
+        if (repository.containsKey(id)) {
+            if (AuthorizedUser.id() == repository.get(id).getUserId()){
+                log.info("delete {}", id);
+                repository.remove(id);
+                return true;
+            } else {
+                log.info("Access violation. Unsuccessful delete {} for", id);
+                return false;
+            }
+
+        } else {
+            log.info("Id not found. Unsuccessful delete {} for", id);
+            return false;
+        }
     }
 
+    // null if not found  or don't have permission
     @Override
     public Meal get(int id) {
-        return repository.get(id);
+        Meal meal = repository.get(id);
+        if ((meal!=null) && (AuthorizedUser.id() == meal.getUserId())){
+            return meal;
+        }
+        return null;
     }
 
     @Override
     public Collection<Meal> getAll() {
-        return repository.values();
+        List<Meal> result;
+        result = repository.values().stream().filter((meal -> (meal.getUserId() == AuthorizedUser.id())))
+                .sorted(Comparator.comparing(Meal::getDateTime))
+                .collect(Collectors.toList());
+        return result;
     }
 }
